Consolidates Gradle task configuration.
