+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/DockerContextTask.java
@@ -22,39 +22,15 @@ import com.google.common.base.Preconditions;
 import com.google.common.io.InsecureRecursiveDeleteException;
 import java.io.IOException;
 import java.nio.file.Paths;
-import java.util.List;
 import javax.annotation.Nullable;
-import org.gradle.api.DefaultTask;
 import org.gradle.api.GradleException;
-import org.gradle.api.tasks.Input;
 import org.gradle.api.tasks.OutputDirectory;
 import org.gradle.api.tasks.TaskAction;
 import org.gradle.api.tasks.options.Option;
 
-public class DockerContextTask extends DefaultTask {
+public class DockerContextTask extends TaskConfiguration {
 
   @Nullable private String targetDir;
-  @Nullable private String baseImage;
-  @Nullable private List<String> jvmFlags;
-  @Nullable private String mainClass;
-
-  @Input
-  @Nullable
-  public String getBaseImage() {
-    return baseImage;
-  }
-
-  @Input
-  @Nullable
-  public List<String> getJvmFlags() {
-    return jvmFlags;
-  }
-
-  @Input
-  @Nullable
-  public String getMainClass() {
-    return mainClass;
-  }
 
   /** The output directory for the Docker context is by default {@code build/jib-dockercontext}. */
   @OutputDirectory
@@ -74,14 +50,14 @@ public class DockerContextTask extends DefaultTask {
   @TaskAction
   public void generateDockerContext() {
     // Asserts required parameters are not null.
-    Preconditions.checkNotNull(baseImage);
-    Preconditions.checkNotNull(jvmFlags);
-    Preconditions.checkNotNull(mainClass);
+    Preconditions.checkNotNull(getFrom());
+    Preconditions.checkNotNull(getFrom().getImage());
+    Preconditions.checkNotNull(getJvmFlags());
 
     // TODO: Refactor with BuildImageTask.
     ProjectProperties projectProperties = new ProjectProperties(getProject(), getLogger());
 
-    String mainClass = this.mainClass;
+    String mainClass = getMainClass();
     if (mainClass == null) {
       mainClass = projectProperties.getMainClassFromJarTask();
       if (mainClass == null) {
@@ -96,8 +72,8 @@ public class DockerContextTask extends DefaultTask {
 
     try {
       new DockerContextGenerator(projectProperties.getSourceFilesConfiguration())
-          .setBaseImage(baseImage)
-          .setJvmFlags(jvmFlags)
+          .setBaseImage(getFrom().getImage())
+          .setJvmFlags(getJvmFlags())
           .setMainClass(mainClass)
           .generate(Paths.get(targetDir));
 
@@ -114,16 +90,6 @@ public class DockerContextTask extends DefaultTask {
       throwMojoExecutionExceptionWithHelpMessage(
           ex, "check if the command-line option `--jib.dockerDir` is set correctly");
     }
-  }
-
-  /**
-   * Applies the configuration from {@code jibExtension}. This must be called before {@link
-   * #generateDockerContext}.
-   */
-  void applyExtension(JibExtension jibExtension) {
-    baseImage = jibExtension.getFrom().getImage();
-    jvmFlags = jibExtension.getJvmFlags();
-    mainClass = jibExtension.getMainClass();
   }
 
   private <T extends Throwable> void throwMojoExecutionExceptionWithHelpMessage(
